pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "xlab"
version = "0.0.2"
authors = [
  { name="Rob Reilink" },
]
maintainers = [
  { name="Robert Roos", email="robert.roos@demcon.com"}
]
description = "Xlab is a Python module that facilitates performing lab experiments"
readme = "README.md"
requires-python = ">=3.7"
dependencies = [
  "pyvisa==1.8.0",
  "numpy",
  "PySide2",
  "pyserial",
]
classifiers = [
"Programming Language :: Python :: 3",
"License :: OSI Approved :: MIT License",
"Operating System :: OS Independent",
]

[options]
package_dir = "src"

[options.packages.find]
where = "src"

[project.urls]
"Homepage" = "https://github.com/DEMCON/python-xlab"
"Documentation" = "https://demcon.github.io/python-xlab"
"Bug Tracker" = "https://github.com/DEMCON/python-xlab/issues"
